requirements.txt
opencv-python
pyzbar
pyautogui
winsound
Pillow
tk
